Update AutoRefreshServerDataWidget.js
Change autorefresh widget to return children prop if it exists. This way, it can wrap any element. I think this will allow the widget's componentDidMount to be called.  That lifecycle function enacts a refresh check.  If the widget doesn't have children and returns null, then it will only check refresh on the constructor call.